+++ app/upgrades/v20/upgrades.go
@@ -12,6 +12,8 @@ import (
 	cltypes "github.com/osmosis-labs/osmosis/v19/x/concentrated-liquidity/types"
 	incentivestypes "github.com/osmosis-labs/osmosis/v19/x/incentives/types"
 	lockuptypes "github.com/osmosis-labs/osmosis/v19/x/lockup/types"
+	poolincenitvestypes "github.com/osmosis-labs/osmosis/v19/x/pool-incentives/types"
+	poolmanagertypes "github.com/osmosis-labs/osmosis/v19/x/poolmanager/types"
 )
 
 type IncentivizedCFMMDirectWhenMigrationLinkPresentError struct {
@@ -19,6 +21,8 @@ type IncentivizedCFMMDirectWhenMigrationLinkPresentError struct {
 	ConcentratedPoolID uint64
 	CFMMGaugeID        uint64
 }
+
+var emptySlice = []string{}
 
 func (e IncentivizedCFMMDirectWhenMigrationLinkPresentError) Error() string {
 	return fmt.Sprintf("CFMM gauge ID (%d) incentivized CFMM pool (%d) directly when migration link is present with concentrated pool (%d)", e.CFMMGaugeID, e.CFMMPoolID, e.ConcentratedPoolID)
@@ -39,11 +43,14 @@ func CreateUpgradeHandler(
 		}
 
 		// Initialize the newly created param
-		keepers.ConcentratedLiquidityKeeper.SetParam(ctx, cltypes.KeyUnrestrictedPoolCreatorWhitelist, []string{})
+		keepers.ConcentratedLiquidityKeeper.SetParam(ctx, cltypes.KeyUnrestrictedPoolCreatorWhitelist, emptySlice)
 
 		// Initialize the new params in incentives for group creation.
 		keepers.IncentivesKeeper.SetParam(ctx, incentivestypes.KeyGroupCreationFee, incentivestypes.DefaultGroupCreationFee)
-		keepers.IncentivesKeeper.SetParam(ctx, incentivestypes.KeyCreatorWhitelist, []string{})
+		keepers.IncentivesKeeper.SetParam(ctx, incentivestypes.KeyCreatorWhitelist, emptySlice)
+
+		// Initialize new param in the poolmanager module with a whitelist allowing to bypass taker fees.
+		keepers.PoolManagerKeeper.SetParam(ctx, poolmanagertypes.KeyReducedTakerFeeByWhitelist, emptySlice)
 
 		// Converts pool incentive distribution records from concentrated gauges to group gauges.
 		err = createGroupsForIncentivePairs(ctx, keepers)
@@ -90,6 +97,12 @@ func createGroupsForIncentivePairs(ctx sdk.Context, keepers *keepers.AppKeepers)
 	// expect any stableswap pool to be linked to concentrated.
 	for i, distrRecord := range distrInfo.Records {
 		gaugeID := distrRecord.GaugeId
+
+		// Gauge with ID zero goes to community pool.
+		if gaugeID == poolincenitvestypes.CommunityPoolDistributionGaugeID {
+			continue
+		}
+
 		gauge, err := keepers.IncentivesKeeper.GetGaugeByID(ctx, gaugeID)
 		if err != nil {
 			return err